Extract 9 defaults-discovery helpers into modules/output_defaults.py
Third refactor pass at modules/output.py.  Moved the entire family of
helpers that compute default values from the quickstart JSON schemas
and prune actual template_variables against those defaults, so emitted
YAML only carries non-default overrides:

  * _values_match                 -- default-equality with type coercion
  * _infer_default_from_options   -- pull first choice from a select
  * _default_from_var             -- {default: X} > toggle-false > select-first
  * _extract_template_defaults    -- run _default_from_var over dict/list
  * _build_collection_defaults    -- crawl quickstart_collections.json
  * _extract_offset_defaults      -- read overlay offset schema
  * _build_overlay_defaults       -- crawl quickstart_overlays.json
  * _build_attribute_defaults     -- crawl quickstart_attributes.json
  * _prune_template_variables     -- drop keys whose value == default

Dependencies: modules.helpers (for load_quickstart_config /
load_quickstart_overlay_config / ts_log) and modules.output_values (for
_normalize_template_value / _coerce_bool / _to_number, from PR #1445).
No Flask, no persistence, no database.

Public surface: output.py re-imports the nine names with '# noqa: F401'
so historical call sites like output._build_overlay_defaults(...) keep
working if any tests reach for them.

Impact:
- modules/output.py: 3538 -> 3343 (-195 / -5.5%)
- modules/output_defaults.py: 241 lines new
- Zero behavior change

Cumulative sprint progress on output.py:
- Start:         3833
- After #1445:   3670  (-163)
- After #1446:   3538  (-132)
- After this:    3343  (-195)  -- total -490 / -12.8%

All 1077 tests pass. Ruff + black clean.
